feat(privacy banner): fix margin, move cookie logic to banner
